Camera now follows the player as expected
Was having design issues as to make the camera follow the player or be isometric. For simplicity sake i have made it just follow the player, but took me a while to get offset and the ability to move the camera along its position and rotation to keep track of the player at all times from the back.