Added win and lose states through singleton
with the singleton design pattern i have set up a game manager that allows us to call the onwin and onlose functions when we need them too. This will be the check for either crashing into the guards (lose) or reaching your waypoint. I want to add a little bit more depth to the game so i will add jumping, which will allow the player to try to find the shortest path possible, but without getting damaged as landing on the guard or outside of the track will be instant death. Risk vs reward.