Happy Progress - Random waypoint algorithm
created a random waypoint algorithm that takes in prefabs and then assigns them to a transform location. At first tried to tackle this by just using the local prefabs transform, but was having issues getting them placed correctly. This is a bandaid fix which is more cubersome but gets the job done